fix mistake

+++ chatunit/ChatServer.cpp
@@ -34,6 +34,7 @@ void ChatServer::inputConnecting() {
 //    socket->flush();
 //    socket->waitForBytesWritten(3000);
 //    socket->close();
+
 }
 
 
@@ -74,7 +75,7 @@ void ChatServer::sendToClient(QString xmldata, QTcpSocket *tcpSocket) {
     out.setVersion(QDataStream::Qt_5_11);
     out << quint32(0) << xmldata;
     out.device()->seek(0);
-    out << quint16(arrBlock.size() - sizeof(quint32));
+    out << quint32(arrBlock.size() - sizeof(quint32));
 
     tcpSocket->write(arrBlock);
 }
